Refactor: split functions from catan.h
