+++ catan.h
@@ -385,766 +385,9 @@ void print_player(mapInfo *map){
     sleep(2);
 }
 
-// SDL implement
-void render_pieces(SDL_Renderer *renderer, mapInfo *map){
-    //set background color
-    SDL_SetRenderDrawColor(renderer, 130, 240, 255, 255);
-    SDL_RenderClear(renderer);
 
-    piece **pieces = map->pieces;
 
-    for(int i = 0; i < PIECE_NUM; i++){
-        int x = pieces[i]->p.x;
-        int y = pieces[i]->p.y;
-        x *= 2;
-        y = y + y / 2 - 1;
-        x *= 20;
-        y *= 20; 
 
-        SDL_Surface *piece_surface = NULL;
-        if(pieces[i]->eco_type == HILL)
-            piece_surface = IMG_Load("picture/hill.png");
-        else if(pieces[i]->eco_type == MOUNTAIN)
-            piece_surface = IMG_Load("picture/mountain.png");
-        else if(pieces[i]->eco_type == FIELD)
-            piece_surface = IMG_Load("picture/field.png");
-        else if(pieces[i]->eco_type == FOREST)
-            piece_surface = IMG_Load("picture/forest.png");
-        else if(pieces[i]->eco_type == PASTURE)
-            piece_surface = IMG_Load("picture/pasture.png");
-        else if(pieces[i]->eco_type == DESERT)
-            piece_surface = IMG_Load("picture/desert.png");
-        else
-            continue;
-        SDL_Texture *piece_texture = SDL_CreateTextureFromSurface(renderer, piece_surface);
-        SDL_Rect piece_rect = {x, y, PIECE_SIZE, PIECE_SIZE};
-        SDL_RenderCopy(renderer, piece_texture, NULL, &piece_rect);
-        SDL_DestroyTexture(piece_texture);
-        SDL_FreeSurface(piece_surface);
-
-        SDL_Surface *num_surface = NULL;
-        if(pieces[i]->number == 2)
-            num_surface = IMG_Load("picture/2.png");
-        else if(pieces[i]->number == 3)
-            num_surface = IMG_Load("picture/3.png");
-        else if(pieces[i]->number == 4)
-            num_surface = IMG_Load("picture/4.png");
-        else if(pieces[i]->number == 5)
-            num_surface = IMG_Load("picture/5.png");
-        else if(pieces[i]->number == 6)
-            num_surface = IMG_Load("picture/6.png");
-        else if(pieces[i]->number == 8)
-            num_surface = IMG_Load("picture/8.png");
-        else if(pieces[i]->number == 9)
-            num_surface = IMG_Load("picture/9.png");
-        else if(pieces[i]->number == 10)
-            num_surface = IMG_Load("picture/10.png");
-        else if(pieces[i]->number == 11)
-            num_surface = IMG_Load("picture/11.png");
-        else if(pieces[i]->number == 12)
-            num_surface = IMG_Load("picture/12.png");
-
-        SDL_Texture *num_texture = SDL_CreateTextureFromSurface(renderer, num_surface);
-        SDL_Rect num_rect = {x + NUM_OFFSET, y + NUM_OFFSET, NUM_SIZE, NUM_SIZE};
-        SDL_RenderCopy(renderer, num_texture, NULL, &num_rect);
-        SDL_DestroyTexture(num_texture);
-        SDL_FreeSurface(num_surface);   
-    }
-
-    SDL_RenderPresent(renderer);
-}
-
-void render_map(SDL_Renderer *renderer, mapInfo *map){
-    //set background color
-    SDL_SetRenderDrawColor(renderer, 0, 206, 255, 255);
-    SDL_RenderClear(renderer);
-
-    if(season_flag){
-        SDL_Surface *piece_surface = NULL;
-
-        if(season_turn == SPRING)
-            piece_surface = IMG_Load("picture/season/spring.png");
-        else if(season_turn == SUMMER)
-            piece_surface = IMG_Load("picture/season/summer.png");
-        else if(season_turn == AUTUMN)
-            piece_surface = IMG_Load("picture/season/autumn.png");
-        else if(season_turn == WINTER)
-            piece_surface = IMG_Load("picture/season/winter.png");
-
-        SDL_Texture *piece_texture = SDL_CreateTextureFromSurface(renderer, piece_surface);
-        SDL_Rect piece_rect = {600, 10, 100, 100};
-        SDL_RenderCopy(renderer, piece_texture, NULL, &piece_rect);
-        SDL_DestroyTexture(piece_texture);
-        SDL_FreeSurface(piece_surface);
-    }
-
-    piece **pieces = map->pieces;
-    landbetween **lands = map->lands;
-    road **roads = map->roads;
-
-    for(int i = 0; i < PIECE_NUM; i++){
-        int x = pieces[i]->p.x;
-        int y = pieces[i]->p.y;
-        x *= 2;
-        y = y + y / 2 - 1;
-        x *= VIC_LEN;
-        y *= VIC_LEN; 
-
-        SDL_Surface *piece_surface = NULL;
-        if(pieces[i]->eco_type == HILL)
-            piece_surface = IMG_Load("picture/hill.png");
-        else if(pieces[i]->eco_type == MOUNTAIN)
-            piece_surface = IMG_Load("picture/mountain.png");
-        else if(pieces[i]->eco_type == FIELD)
-            piece_surface = IMG_Load("picture/field.png");
-        else if(pieces[i]->eco_type == FOREST)
-            piece_surface = IMG_Load("picture/forest.png");
-        else if(pieces[i]->eco_type == PASTURE)
-            piece_surface = IMG_Load("picture/pasture.png");
-        else if(pieces[i]->eco_type == DESERT)
-            piece_surface = IMG_Load("picture/desert.png");
-        else
-            continue;
-        SDL_Texture *piece_texture = SDL_CreateTextureFromSurface(renderer, piece_surface);
-        SDL_Rect piece_rect = {x, y, PIECE_SIZE, PIECE_SIZE};
-        SDL_RenderCopy(renderer, piece_texture, NULL, &piece_rect);
-        SDL_DestroyTexture(piece_texture);
-        SDL_FreeSurface(piece_surface);
-
-        SDL_Surface *num_surface = NULL;
-        SDL_Surface *rob_surface = NULL;
-        if(pieces[i]->number == 2)
-            num_surface = IMG_Load("picture/2.png");
-        else if(pieces[i]->number == 3)
-            num_surface = IMG_Load("picture/3.png");
-        else if(pieces[i]->number == 4)
-            num_surface = IMG_Load("picture/4.png");
-        else if(pieces[i]->number == 5)
-            num_surface = IMG_Load("picture/5.png");
-        else if(pieces[i]->number == 6)
-            num_surface = IMG_Load("picture/6.png");
-        else if(pieces[i]->number == 8)
-            num_surface = IMG_Load("picture/8.png");
-        else if(pieces[i]->number == 9)
-            num_surface = IMG_Load("picture/9.png");
-        else if(pieces[i]->number == 10)
-            num_surface = IMG_Load("picture/10.png");
-        else if(pieces[i]->number == 11)
-            num_surface = IMG_Load("picture/11.png");
-        else if(pieces[i]->number == 12)
-            num_surface = IMG_Load("picture/12.png");
-        
-        SDL_Texture *num_texture = SDL_CreateTextureFromSurface(renderer, num_surface);
-        SDL_Rect num_rect = {x + NUM_OFFSET, y + NUM_OFFSET, NUM_SIZE, NUM_SIZE};
-        SDL_RenderCopy(renderer, num_texture, NULL, &num_rect);
-        SDL_DestroyTexture(num_texture);
-
-        
-        if(pieces[i]->robberFlag == true){
-            rob_surface = IMG_Load("picture/CreeperFace.png");
-            SDL_Texture *rob_texture = SDL_CreateTextureFromSurface(renderer, rob_surface);
-            SDL_Rect rob_rect = {x + NUM_OFFSET + PIECE_SIZE / 4, y + NUM_OFFSET, NUM_SIZE - PIECE_SIZE / 10, NUM_SIZE - PIECE_SIZE / 10};
-            SDL_RenderCopy(renderer, rob_texture, NULL, &rob_rect);
-            SDL_DestroyTexture(rob_texture);
-            SDL_FreeSurface(rob_surface);
-        }   
-        
-
-        SDL_FreeSurface(num_surface);   
-    }
-
-    for(int i = 0; i < ROAD_NUM; i++){
-        int x, y;
-
-        SDL_Surface *road_surface = NULL;
-        if(roads[i]->dir == LD){
-            if(roads[i]->owner == NOBODY)
-                road_surface = IMG_Load("picture/LD_black.png");
-            else if(roads[i]->owner == PLAYER1)
-                road_surface = IMG_Load("picture/LD_red.png");
-            else if(roads[i]->owner == PLAYER2)
-                road_surface = IMG_Load("picture/LD_white.png");
-            else if(roads[i]->owner == PLAYER3)
-                road_surface = IMG_Load("picture/LD_orange.png");
-            else if(roads[i]->owner == PLAYER4)
-                road_surface = IMG_Load("picture/LD_blue.png");
-
-            x = roads[i]->start.x;
-            x = x * 2 * VIC_LEN;
-            y = roads[i]->start.y;
-            y = y + y / 2;
-            y = y * VIC_LEN;
-            y = y + PIECE_SIZE / 2 - PIECE_SIZE / 40;
-
-            double angle = 45 - 26.5650511770;
-            SDL_Point center = {PIECE_SIZE / 2, 0};            
-            SDL_Rect dstRect = {x, y, PIECE_SIZE / 2, PIECE_SIZE / 2};
-
-            SDL_Texture *road_texture = SDL_CreateTextureFromSurface(renderer, road_surface);
-            // SDL_Rect road_rect = {x, y, PIECE_SIZE / 2, PIECE_SIZE / 2};
-            // SDL_RenderCopy(renderer, road_texture, NULL, &road_rect);
-            SDL_RenderCopyEx(renderer, road_texture, NULL, &dstRect, angle, &center, SDL_FLIP_NONE);
-            SDL_DestroyTexture(road_texture);
-        }
-        else if(roads[i]->dir == RD){
-            if(roads[i]->owner == NOBODY)
-                road_surface = IMG_Load("picture/RD_black.png");
-            else if(roads[i]->owner == PLAYER1)
-                road_surface = IMG_Load("picture/RD_red.png");
-            else if(roads[i]->owner == PLAYER2)
-                road_surface = IMG_Load("picture/RD_white.png");
-            else if(roads[i]->owner == PLAYER3)
-                road_surface = IMG_Load("picture/RD_orange.png");
-            else if(roads[i]->owner == PLAYER4)
-                road_surface = IMG_Load("picture/RD_blue.png");
-
-            x = roads[i]->start.x;
-            x = x * 2 * VIC_LEN;
-            x += PIECE_SIZE / 2 + PIECE_SIZE / 10 + PIECE_SIZE / 40;
-            y = roads[i]->start.y;
-            y = y + y / 2;
-            y = y * VIC_LEN;
-            y = y + PIECE_SIZE / 2 - PIECE_SIZE / 40;
-
-            double angle = -1 * (45 - 26.5650511770);
-            SDL_Point center = {0, PIECE_SIZE / 2};            
-            SDL_Rect dstRect = {x, y, PIECE_SIZE / 2, PIECE_SIZE / 2};
-
-            SDL_Texture *road_texture = SDL_CreateTextureFromSurface(renderer, road_surface);
-            // SDL_Rect road_rect = {x, y, PIECE_SIZE / 2, PIECE_SIZE / 2};
-            // SDL_RenderCopy(renderer, road_texture, NULL, &road_rect);
-            SDL_RenderCopyEx(renderer, road_texture, NULL, &dstRect, angle, &center, SDL_FLIP_NONE);
-            SDL_DestroyTexture(road_texture);
-        }
-        else if(roads[i]->dir == D){
-            if(roads[i]->owner == NOBODY)
-                road_surface = IMG_Load("picture/D_black.png");
-            else if(roads[i]->owner == PLAYER1)
-                road_surface = IMG_Load("picture/D_red.png");
-            else if(roads[i]->owner == PLAYER2)
-                road_surface = IMG_Load("picture/D_white.png");
-            else if(roads[i]->owner == PLAYER3)
-                road_surface = IMG_Load("picture/D_orange.png");
-            else if(roads[i]->owner == PLAYER4)
-                road_surface = IMG_Load("picture/D_blue.png");
-
-            x = roads[i]->start.x;
-            x = x * 2 * VIC_LEN;
-            x += PIECE_SIZE / 4;
-            y = roads[i]->start.y;
-            y = y + y / 2;
-            y = y * VIC_LEN;
-            y = y + PIECE_SIZE / 2 + PIECE_SIZE / 20;
-
-
-            SDL_Texture *land_texture = SDL_CreateTextureFromSurface(renderer, road_surface);
-            SDL_Rect land_rect = {x, y, PIECE_SIZE / 2, PIECE_SIZE / 2};
-            SDL_RenderCopy(renderer, land_texture, NULL, &land_rect);
-            SDL_DestroyTexture(land_texture);
-        }
-        SDL_FreeSurface(road_surface);
-    }
-
-    for(int i = 0; i < LAND_NUM; i++){
-        int x = lands[i]->p.x;
-        int y = lands[i]->p.y;
-        x *= 2;
-        y = y + y / 2;
-        x *= VIC_LEN;
-        y *= VIC_LEN; 
-        x = x + PIECE_SIZE / 2 - LAND_SIZE / 2 + CIRCLE_RAD;
-        y = y + PIECE_SIZE / 2 - LAND_SIZE / 2 + CIRCLE_RAD;
-
-        if(valid_point_mat[lands[i]->p.y][lands[i]->p.x] == TYPE_PORT_ANY)
-            render_circle(renderer, x, y, CIRCLE_RAD, 255, 80, 255, ALPHA);
-        else if(valid_point_mat[lands[i]->p.y][lands[i]->p.x] == TYPE_PORT_BRICK)
-            render_circle(renderer, x, y, CIRCLE_RAD, 255, 86, 80, ALPHA);
-        else if(valid_point_mat[lands[i]->p.y][lands[i]->p.x] == TYPE_PORT_LUMBER)
-            render_circle(renderer, x, y, CIRCLE_RAD, 200, 98, 0, ALPHA);
-        else if(valid_point_mat[lands[i]->p.y][lands[i]->p.x] == TYPE_PORT_WOOL)
-            render_circle(renderer, x, y, CIRCLE_RAD, 210, 210, 210, ALPHA);
-        else if(valid_point_mat[lands[i]->p.y][lands[i]->p.x] == TYPE_PORT_GRAIN)
-            render_circle(renderer, x, y, CIRCLE_RAD, 246, 218, 0, ALPHA);
-        else if(valid_point_mat[lands[i]->p.y][lands[i]->p.x] == TYPE_PORT_ORE)
-            render_circle(renderer, x, y, CIRCLE_RAD, 104, 78, 56, ALPHA);
-    }   
-
-
-    for(int i = 0; i < LAND_NUM; i++){
-        int x = lands[i]->p.x;
-        int y = lands[i]->p.y;
-        x *= 2;
-        y = y + y / 2;
-        x *= VIC_LEN;
-        y *= VIC_LEN; 
-
-        SDL_Surface *land_surface = NULL;
-        if(lands[i]->has_building){
-            int own = lands[i]->owner;
-            int b= lands[i]->building;
-
-            if(own == PLAYER1){
-                if(b == SETTLEMENT)
-                    land_surface = IMG_Load("picture/red_settlement.png");
-                else if(b == CITY)
-                    land_surface = IMG_Load("picture/red_city.png");
-            }
-            else if(own == PLAYER2){
-                if(b == SETTLEMENT)
-                    land_surface = IMG_Load("picture/white_settlement.png");
-                else if(b == CITY)
-                    land_surface = IMG_Load("picture/white_city.png");
-            }
-            else if(own == PLAYER3){
-                if(b == SETTLEMENT)
-                    land_surface = IMG_Load("picture/orange_settlement.png");
-                else if(b == CITY)
-                    land_surface = IMG_Load("picture/orange_city.png");
-            }
-            else if(own == PLAYER4){
-                if(b == SETTLEMENT)
-                    land_surface = IMG_Load("picture/blue_settlement.png");
-                else if(b == CITY)
-                    land_surface = IMG_Load("picture/blue_city.png");
-            }
-        }
-
-        x = x + PIECE_SIZE / 2 - LAND_SIZE / 2;
-        y = y + PIECE_SIZE / 2 - LAND_SIZE / 2;
-        SDL_Texture *land_texture = SDL_CreateTextureFromSurface(renderer, land_surface);
-        SDL_Rect land_rect = {x, y, LAND_SIZE, LAND_SIZE};
-        SDL_RenderCopy(renderer, land_texture, NULL, &land_rect);
-        SDL_DestroyTexture(land_texture);
-        SDL_FreeSurface(land_surface);
-    }   
-
-    SDL_RenderPresent(renderer);
-}
-
-void render_circle(SDL_Renderer *renderer, int x, int y, int rad, int r, int g, int b, int a){
-    //setting color 
-    SDL_SetRenderDrawBlendMode(renderer, SDL_BLENDMODE_BLEND);
-    SDL_SetRenderDrawColor(renderer, r, g, b, a);
-
-    // draw circle and fill it with the color
-    for(int i = 0; i < 360; i++){
-        double angle = i * 3.1415926535 / 180;
-        int x1 = x + rad * cos(angle);
-        int y1 = y + rad * sin(angle);
-        SDL_RenderDrawLine(renderer, x1, y1, x, y);
-    }
-
-    SDL_RenderCopy(renderer, NULL, NULL, NULL);
-}
-
-
-// return the length of a list, excluding the head
-int list_head_length(struct list_head *head){
-    int length = 0;
-    struct list_head *pos = head->next;
-    while(pos != head){
-        length++;
-        pos = pos->next;
-    }
-    return length;
-}
-
-// return the pointer of the node at th position, excluding the head
-struct list_head *list_head_at(struct list_head *head, int th){
-    if(th <= 0 || th > list_head_length(head))
-        return NULL;
-    
-    struct list_head *pos = head;
-    for(int i = 0; i < th; i++)
-        pos = pos->next;
-    return pos;
-}
-
-void randomize(void **array, size_t n, size_t size){
-    srand(time(NULL));
-    if(n > 1){
-        for(int i = 0; i < TIMES; i++){
-            int a = rand() % n;
-            int b = rand() % n;
-            if(a == b)
-                continue;
-
-            void *temp = malloc(size);
-            memcpy(temp, array[a], size);
-            memcpy(array[a], array[b], size);
-            memcpy(array[b], temp, size);
-            free(temp);
-        }
-    }
-}
-
-void init_game(piece ***pieces, landbetween ***lands, road ***roads, player ***players, struct list_head **devcards){
-    // init item
-    *pieces = init_piece();
-    *players = init_player();
-    *lands = init_landbetween();
-    *roads = init_road();   
-    *devcards = init_devcard();
-}
-
-player **init_player(){
-    player **players = (player **)malloc(sizeof(player *) * PLAYER_NUM);
-    for(int i = 0; i < PLAYER_NUM; i++){
-        players[i] = (player *)malloc(sizeof(player));
-        players[i]->id = i + 1;
-        players[i]->VP = 0;
-        for(int j = 0; j < 5; j++)
-            players[i]->resource[j] = 0;
-        for(int j = 0; j < 3; j++)
-            players[i]->number_of_building[j] = 0;
-        players[i]->number_of_knights = 0;
-        players[i]->length_of_road = 0;
-        players[i]->number_of_dev_card = 0;
-        players[i]->has_longest_road = false;
-        players[i]->has_most_knights = false;
-        struct list_head *devcard_list = (struct list_head *)malloc(sizeof(struct list_head));
-        devcard_list->next = devcard_list;
-        devcard_list->prev = devcard_list;
-        players[i]->devcard_list = devcard_list;
-    }
-
-    randomize((void **)players, PLAYER_NUM, sizeof(player *));
-    return players;
-}
-
-piece **init_piece(){
-    srand(time(NULL));
-    piece **pieces = (piece **)malloc(sizeof(piece *) * PIECE_NUM);
-    int p = 0;
-    point desert = {0, 0};
-    for(int i = 0; i < Y_LONG; i++){
-        for(int j = 0; j < X_LONG; j++){
-            if(valid_point_mat[i][j] == TYPE_PIECE){
-                pieces[p] = (piece *)malloc(sizeof(piece));
-                pieces[p]->p.x = j;
-                pieces[p]->p.y = i;
-                while(1){
-                    int eco_type = rand() % 6;
-                    if(eco_num[eco_type] > 0){
-                        if(eco_type == DESERT){
-                            pieces[p]->robberFlag = true;
-                            pieces[p]->number = 0;
-                        }
-                        else
-                            pieces[p]->robberFlag = false;
-                        pieces[p]->eco_type = eco_type;
-                        eco_num[eco_type]--;
-                        break;
-                    }
-                }
-                p++;
-            }
-        }
-    }
-
-    randomize((void **)pieces, PIECE_NUM, sizeof(piece *));
-
-    for(int i = 0; i < PIECE_NUM; i++){
-        if(pieces[i]->eco_type == DESERT){
-            desert.x = pieces[i]->p.x;
-            desert.y = pieces[i]->p.y;
-            pieces[i]->number = 7;
-            break;
-        }
-    }
-
-    const point pos[PIECE_NUM] = {
-        {1, 6}, {2, 8}, {3, 10}, {5, 10}, {7, 10}, {8, 8},
-        {9, 6}, {8, 4}, {7, 2}, {5, 2}, {3, 2}, {2, 4},
-        {3, 6}, {4, 8}, {6, 8}, {7, 6}, {6, 4}, {4, 4}, {5, 6},
-    };
-
-    const int val[PIECE_NUM-1] = {5, 2, 6, 3, 8, 10, 9, 12, 11, 4, 8, 10, 9, 4, 5, 6, 3, 11};
-
-    bool desertFlag = false;
-    for(int i = 0; i < PIECE_NUM; i++){
-        if(pos[i].x == desert.x && pos[i].y == desert.y){
-            desertFlag = true;
-            continue;
-        }
-        for(int j = 0; j < PIECE_NUM; j++){
-            if(pieces[j]->p.x == pos[i].x && pieces[j]->p.y == pos[i].y){
-                if(desertFlag)
-                    pieces[j]->number = val[i-1];
-                else
-                    pieces[j]->number = val[i];
-                pieces[j]->robberFlag = false;
-                break;
-            }
-        }
-    }
-
-    eco_num[DESERT] = 1;
-    eco_num[HILL] = 3;
-    eco_num[FOREST] = 4;
-    eco_num[MOUNTAIN] = 3;
-    eco_num[FIELD] = 4;
-    eco_num[PASTURE] = 4;
-
-    return pieces;
-}
-
-int piece_index(int x, int y, piece **pieces){
-    for(int i = 0; i < PIECE_NUM; i++){
-        if(pieces[i]->p.x == x && pieces[i]->p.y == y)
-            return i;
-    }
-    return -1;
-}
-
-int land_index(const int x, const int y, landbetween **lands){
-    for(int i = 0; i < LAND_NUM; i++){
-        if(lands[i]->p.x == x && lands[i]->p.y == y)
-            return i;
-    }
-    return -1;
-}
-
-int player_index(const int id, player **players){
-    for(int i = 0; i < PLAYER_NUM; i++){
-        if(players[i]->id == id)
-            return i;
-    }
-    return -1;
-}
-
-int point_type(const int x, const int y){
-    for(int i = 0; i < Y_LONG; i++){
-        for(int j = 0; j < X_LONG; j++){
-            if(valid_point_mat[i][j] == TYPE_PIECE && x == j && y == i)
-                return TYPE_PIECE;
-            else if(valid_point_mat[i][j] == TYPE_LANDBETWEEN && x == j && y == i)
-                return TYPE_LANDBETWEEN;
-            else if(valid_point_mat[i][j] == TYPE_PORT_ANY && x == j && y == i)
-                return TYPE_PORT_ANY;
-            else if(valid_point_mat[i][j] == TYPE_PORT_BRICK && x == j && y == i)
-                return TYPE_PORT_BRICK;
-            else if(valid_point_mat[i][j] == TYPE_PORT_LUMBER && x == j && y == i)
-                return TYPE_PORT_LUMBER;
-            else if(valid_point_mat[i][j] == TYPE_PORT_WOOL && x == j && y == i)
-                return TYPE_PORT_WOOL;
-            else if(valid_point_mat[i][j] == TYPE_PORT_GRAIN && x == j && y == i)
-                return TYPE_PORT_GRAIN;
-            else if(valid_point_mat[i][j] == TYPE_PORT_ORE && x == j && y == i)
-                return TYPE_PORT_ORE;
-        }
-    }
-    return -1;
-}
-
-int resource_num(const int id, mapInfo *map){
-    player **players = map->players;
-    int index = player_index(id, players);
-    int sum = 0;
-    for(int i = 0; i < 5; i++)
-        sum += players[index]->resource[i];
-    return sum;
-}
-
-landbetween **init_landbetween(){
-    landbetween **lands = (landbetween **)malloc(sizeof(landbetween *) * LAND_NUM);
-    for(int i = 0; i < LAND_NUM; i++){
-        lands[i] = (landbetween *)malloc(sizeof(landbetween));
-        lands[i]->has_building = false;
-        lands[i]->owner = -1;
-        lands[i]->building = -1;
-    }
-
-    int p = 0;
-    for(int i = 0; i < Y_LONG; i++){
-        for(int j = 0; j < X_LONG; j++){
-            if(valid_point_mat[i][j] != TYPE_PIECE && valid_point_mat[i][j] != 0){
-                lands[p]->p.x = j;
-                lands[p]->p.y = i;
-                lands[p]->type = valid_point_mat[i][j];
-                p++;
-            }
-        }
-    }
-
-    return lands;
-}
-
-road **init_road(){
-    road **roads = (road **)malloc(sizeof(road *) * ROAD_NUM);
-    for(int i = 0; i < ROAD_NUM; i++){
-        roads[i] = (road *)malloc(sizeof(road));
-        roads[i]->owner = NOBODY;
-    }
-
-    int p = 0;
-    /*      /\/\/\      */
-    for(int i = 0; i < 3; i++){
-        roads[p]->start.x = 3 + i * 2;
-        roads[p]->start.y = 0;
-        roads[p+1]->start = roads[p]->start;
-        roads[p]->end.x = 2 + i * 2;
-        roads[p]->end.y = 1;
-        roads[p+1]->end.x = 4 + i * 2;
-        roads[p+1]->end.y = 1;
-        roads[p]->dir = LD;
-        roads[p+1]->dir = RD;
-        p += 2;
-    }
-
-    /*      | | | |        */
-    for(int i = 0; i < 4; i++){
-        roads[p]->start.x = 2 + i * 2;
-        roads[p]->start.y = 1;
-        roads[p]->end.x = 2 + i * 2;
-        roads[p]->end.y = 2;
-        roads[p]->dir = D;
-        p++;
-    }
-
-    /*      /\/\/\/\        */
-    for(int i = 0; i < 4; i++){
-        roads[p]->start.x = 2 + i * 2;
-        roads[p]->start.y = 2;
-        roads[p+1]->start = roads[p]->start;
-        roads[p]->end.x = 1 + i * 2;
-        roads[p]->end.y = 3;
-        roads[p+1]->end.x = 3 + i * 2;
-        roads[p+1]->end.y = 3;
-        roads[p]->dir = LD;
-        roads[p+1]->dir = RD;
-        p += 2;
-    }
-
-    /*      | | | | |       */
-    for(int i = 0; i < 5; i++){
-        roads[p]->start.x = 1 + i * 2;
-        roads[p]->start.y = 3;
-        roads[p]->end.x = 1 + i * 2;
-        roads[p]->end.y = 4;
-        roads[p]->dir = D;
-        p++;
-    }
-
-    /*      /\/\/\/\/\      */
-    for(int i = 0; i < 5; i++){
-        roads[p]->start.x = 1 + i * 2;
-        roads[p]->start.y = 4;
-        roads[p+1]->start = roads[p]->start;
-        roads[p]->end.x = 0 + i * 2;
-        roads[p]->end.y = 5;
-        roads[p+1]->end.x = 2 + i * 2;
-        roads[p+1]->end.y = 5;
-        roads[p]->dir = LD;
-        roads[p+1]->dir = RD;
-        p += 2;
-    }
-
-    /*      | | | | | |     */
-    for(int i = 0; i < 6; i++){
-        roads[p]->start.x = i * 2;
-        roads[p]->start.y = 5;
-        roads[p]->end.x = i * 2;
-        roads[p]->end.y = 6;
-        roads[p]->dir = D;
-        p++;
-    }
-
-    /*      \/\/\/\/\/      */
-    for(int i = 0; i < 5; i++){
-        roads[p]->end.x = 1 + i * 2;
-        roads[p]->end.y = 7;
-        roads[p+1]->end = roads[p]->end;
-        roads[p]->start.x = i * 2;
-        roads[p]->start.y = 6;
-        roads[p+1]->start.x = 2 + i * 2;
-        roads[p+1]->start.y = 6;
-        roads[p]->dir = RD;
-        roads[p+1]->dir = LD;
-        p += 2;
-    }
-
-    /*      | | | | |       */
-    for(int i = 0; i < 5; i++){
-        roads[p]->start.x = 1 + i * 2;
-        roads[p]->start.y = 7;
-        roads[p]->end.x = 1 + i * 2;
-        roads[p]->end.y = 8;
-        roads[p]->dir = D;
-        p++;
-    }
-
-    /*      \/\/\/\/        */
-    for(int i = 0; i < 4; i++){
-        roads[p]->end.x = 2 + i * 2;
-        roads[p]->end.y = 9;
-        roads[p+1]->end = roads[p]->end;
-        roads[p]->start.x = 1 + i * 2;
-        roads[p]->start.y = 8;
-        roads[p+1]->start.x = 3 + i * 2;
-        roads[p+1]->start.y = 8;
-        roads[p]->dir = RD;
-        roads[p+1]->dir = LD;
-        p += 2;
-    }
-
-    /*      | | | |     */
-    for(int i = 0; i < 4; i++){
-        roads[p]->start.x = 2 + i * 2;
-        roads[p]->start.y = 9;
-        roads[p]->end.x = 2 + i * 2;
-        roads[p]->end.y = 10;
-        roads[p]->dir = D;
-        p++;
-    }
-
-    /*      \/\/\/      */
-    for(int i = 0; i < 3; i++){
-        roads[p]->end.x = 3 + i * 2;
-        roads[p]->end.y = 11;
-        roads[p+1]->end = roads[p]->end;
-        roads[p]->start.x = 2 + i * 2;
-        roads[p]->start.y = 10;
-        roads[p+1]->start.x = 4 + i * 2;
-        roads[p+1]->start.y = 10;
-        roads[p]->dir = RD;
-        roads[p+1]->dir = LD;
-        p += 2;
-    }
-
-    return roads;
-}
-
-struct list_head *init_devcard(){
-    struct list_head *devcard_list = (struct list_head *)malloc(sizeof(struct list_head));
-    devcard_list->next = devcard_list;
-    devcard_list->prev = devcard_list;
-
-    for(int i = 0; i < KNIGHT_NUM + MONOPOLY_NUM + FREE_ROAD_BUILDING_NUM + YEAR_OF_PLENTY_NUM + VICTORY_POINT_NUM; i++){
-        devcard *card = (devcard *)malloc(sizeof(devcard));
-        if(i < KNIGHT_NUM)
-            card->type = KNIGHT;
-        else if(i < KNIGHT_NUM + MONOPOLY_NUM)
-            card->type = MONOPOLY;
-        else if(i < KNIGHT_NUM + MONOPOLY_NUM + FREE_ROAD_BUILDING_NUM)
-            card->type = FREE_ROAD_BUILDING;
-        else if(i < KNIGHT_NUM + MONOPOLY_NUM + FREE_ROAD_BUILDING_NUM + YEAR_OF_PLENTY_NUM)
-            card->type = YEAR_OF_PLENTY;
-        else
-            card->type = VICTORY_POINT;
-        card->used = -1;
-        list_add_tail(&card->node, devcard_list);
-    }
-
-    srand(time(NULL));
-    for(int i = 0; i < TIMES; i++){
-        int th = rand() % list_head_length(devcard_list) + 1;
-        struct list_head *pos = list_head_at(devcard_list, th);
-        list_move_tail(pos, devcard_list);
-    }
-
-    return devcard_list;
-}
 
 int roll_dice(){
     srand(time(NULL));
