+++ linuxlist.h
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #pragma once
 
 struct list_head
